Created more React components
