+++ shambomon/assets/js/components/hp.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+// Renders an HP bar
+export default function HP(props) {
+  return (
+    <div className="hp-bar" id={props.type}></div>
+  );
+}
