+++ shambomon/assets/js/components/shambomon.jsx
@@ -4,6 +4,8 @@ import { Button } from 'reactstrap';
 
 import Waiting from './waiting';
 import Winner from './winner';
+import Messages from './messages';
+import Battlefield from './battlefield';
 
 export default function start_game(root, channel, user_id) {
   ReactDOM.render(<Shambomon channel={ channel } user={ user_id } />, root);
@@ -134,206 +136,5 @@ class Shambomon extends React.Component {
         </div>
       );
     }
-  }
-}
-
-// Renders the attack history sidebar
-function Messages(props) {
-  let messages = props.state.messages;
-  // Skip if no messages
-  if (messages.length == 0) {
-    return (
-      <div className="col-md-3" id="messages">
-        <h3>ATTACK HISTORY</h3>
-      </div>
-    );
-  }
-  else {
-    var msgs = [];
-    // Make the newest set of messages a different color
-    for (var i = messages.length - 1; i >= messages.length - 3; i--) {
-      msgs.push(<p className="last-messages">>> {messages[i]}</p>);
-    }
-    if (messages.length >= 3) {
-      msgs.push(<br />);
-      for (var i = messages.length - 4; i >= 0; i--) {
-        msgs.push(<p>>> {messages[i]}</p>);
-        // Add a space between each round
-        if (i % 3 == 0) {
-          msgs.push(<br />);
-        }
-      }
-    }
-
-    return (
-      <div className="col-md-3" id="messages">
-        <h3>ATTACK HISTORY</h3>
-        {msgs}
-      </div>
-    );
-  }
-}
-
-// Renders the battlefield, where the user's character is at the bottom
-function Battlefield(props) {
-  let players = props.state.players;
-  var player, opponent;
-  var playerImg, opponentImg;
-  // User is Player 1
-  if (players[0].id == props.id) {
-    player = 0;
-    opponent = 1;
-    playerImg = "/images/" + players[player].char + "-battle.png";
-    opponentImg = "/images/" + players[opponent].char + "-battle.png";
-  }
-  // User is Player 2
-  else {
-    player = 1;
-    opponent = 0;
-    playerImg = "/images/" + players[player].char + "-battle.png";
-    opponentImg = "/images/" + players[opponent].char + "-battle.png";
-  }
-
-  return (
-    <div className="col-md-9" id="arena">
-      {/* All images taken from http://www.ign.com/pokedex/pokemon/ */}
-      {/* Top */}
-      <div className="row player-info">
-        <div className="col-md-9">
-          <PlayerInfo player={opponent} state={props.state} left={false} />
-        </div>
-        <Player img={opponentImg} />
-      </div>
-      <Turn state={props.state} id={props.id} />
-      {/* Bottom */}
-      <div className="row player-info">
-        <Player img={playerImg} />
-        <div className="col-md-9 right-side">
-          <PlayerInfo player={player} state={props.state} left={true} />
-          <Attack attack={props.attack} state={props.state} id={props.id} />
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Renders the message indicating whose turn it is
-function Turn(props) {
-  let players = props.state.players;
-  var msg = "";
-  if (players[props.state.turn].id == props.id) {
-    msg = "YOUR TURN";
-  }
-  else if (props.state.spectators.includes(props.id)) {
-    msg = "YOU ARE SPECTATING";
-  }
-  else {
-    msg = "OPPONENT'S TURN";
-  }
-  return (
-    <div className="waiting">
-      {/* Image credits to http://pixelartmaker.com/art/b73566a14633720 */}
-      <p>
-        <img src="/images/Pokeball.png" />
-        {msg}
-        <img src="/images/Pokeball.png" />
-      </p>
-    </div>
-  );
-}
-
-// Renders the player's character
-function Player(props) {
-  return (
-    <div className="col-md-3">
-      <img src={props.img}/>
-    </div>
-  );
-}
-
-// Renders the player's character name and HP
-function PlayerInfo(props) {
-  var health, name, pokemon, pos;
-  let players = props.state.players;
-  health = players[props.player].health;
-  name = players[props.player].char;
-
-  var hp = []
-  for (var i = 0; i < health; i++) {
-    // Use 'and' because JS is weird about double inequalities
-    if ((50 < health) && (health <= 100)) {
-      hp.push(<HP type={"alive"} />);
-    }
-    else if ((20 < health) && (health <= 50)) {
-      hp.push(<HP type={"low"} />);
-    }
-    else {
-      hp.push(<HP type={"critical"} />);
-    }
-  }
-  for (var i = health; i < 100; i++) {
-    hp.push(<HP type={"dead"} />);
-  }
-
-  // Pad the bottom player's name only
-  if (props.left) {
-    return (
-      <div>
-        <p id="player-name">{name}</p>
-        <div className="hp">
-          {hp}
-          <p>HP {health} / 100</p>
-        </div>
-      </div>
-    );
-  }
-  else {
-    return (
-      <div>
-        <p>{name}</p>
-        <div className="hp">
-          {hp}
-          <p>HP {health} / 100</p>
-        </div>
-      </div>
-    );
-  }
-}
-
-// Renders an HP bar
-function HP(props) {
-  return (
-    <div className="hp-bar" id={props.type}></div>
-  );
-}
-
-// Renders the attack buttons
-function Attack(props) {
-  let players = props.state.players;
-  var disabled = players[props.state.turn].id != props.id;
-
-  // Don't show the attack buttons for spectators
-  if (!props.state.spectators.includes(props.id)) {
-    return (
-      <div className="attack">
-        <span>Choose an attack: </span>
-        {/* Image credits to https://www.dfpeducation.com/play-the-google-game */}
-        <span>
-          <input type="image" src="/images/Rock.png" disabled={disabled}
-            onClick={() => props.attack("Rock")} alt="Rock" />
-        </span>
-        <span>
-          <input type="image" src="/images/Paper.png" disabled={disabled}
-            onClick={() => props.attack("Paper")} alt="Paper" />
-        </span>
-        <span>
-          <input type="image" src="/images/Scissor.png" disabled={disabled}
-            onClick={() => props.attack("Scissor")} alt="Scissor" />
-        </span>
-      </div>
-    );
-  }
-  else {
-    return (<div></div>);
   }
 }
